remove old parser
