+++ crates/yew_router_route_parser/src/parser/core.rs
@@ -1,7 +1,7 @@
 //! Core functions for working with the route parser.
 use crate::parser::CaptureOrExact;
 use crate::parser::RouteParserToken;
-use crate::parser::{Capture, CaptureVariant};
+use crate::parser::{RefCaptureVariant};
 use nom::branch::alt;
 use nom::bytes::complete::{is_not, tag, take};
 use nom::character::complete::char;
@@ -42,119 +42,16 @@ pub fn match_exact(i: &str) -> IResult<&str, RouteParserToken, VerboseError<&str
     context(
         "match",
         map(valid_exact_match_characters, |ident| {
-            RouteParserToken::Exact(ident.to_string())
+            RouteParserToken::Exact(ident)
         }),
     )(i)
 }
 
-/// Matches any of the capture variants
-///
-/// * {name}
-/// * {*:name}
-/// * {5:name}
-pub fn capture(i: &str) -> IResult<&str, RouteParserToken, VerboseError<&str>> {
-    // Capture the variant.
-    let capture_variants = alt((
-        // This can be terminated by either the end of the match section, or by the beginning of a allowed_matches section.
-        map(preceded(tag("*:"), valid_ident_characters), |s| {
-            CaptureVariant::ManyNamed(s.to_string())
-        }),
-        map(valid_ident_characters, |s| {
-            CaptureVariant::Named(s.to_string())
-        }),
-        map(
-            separated_pair(digit1, char(':'), valid_ident_characters),
-            |(n, s)| CaptureVariant::NumberedNamed {
-                sections: n.parse().expect("Should parse digits"),
-                name: s.to_string(),
-            },
-        ),
-    ));
-
-    let allowed_matches = map(
-        separated_list(char('|'), valid_exact_match_characters), // TODO this character set may be too limiting (you may want / to appear in the matched section).
-        |x: Vec<&str>| x.into_iter().map(String::from).collect::<Vec<_>>(),
-    );
-    let allowed_matches = delimited(char('('), allowed_matches, char(')'));
-
-    // Allow capturing the variant, and optionally a list of strings in the form of (string|string|string|...)
-    let capture_inner = map(
-        pair(capture_variants, opt(allowed_matches)),
-        |(cv, allowed_matches): (CaptureVariant, Option<Vec<String>>)| Capture {
-            capture_variant: cv,
-            allowed_captures: allowed_matches,
-        },
-    );
-
-    // wraps the variant and allowed_matches list between `{` and `}`.
-    context(
-        "capture",
-        map(
-            delimited(char('{'), capture_inner, char('}')),
-            RouteParserToken::Capture,
-        ),
-    )(i)
-}
-
-/// Matches either "item" or "{capture}"
-/// It returns a subset enum of Token.
-pub fn capture_or_match(i: &str) -> IResult<&str, CaptureOrExact, VerboseError<&str>> {
-    let (i, token) = context("capture or match", alt((capture, match_exact)))(i)?;
-    let token = match token {
-        RouteParserToken::Capture(variant) => CaptureOrExact::Capture(variant),
-        RouteParserToken::Exact(m) => CaptureOrExact::Exact(m),
-        _ => unreachable!("Only should handle captures and matches"),
-    };
-    Ok((i, token))
-}
 
 #[cfg(test)]
 mod test {
     use super::*;
 
-    #[test]
-    fn capture_named_test() {
-        let cap = capture("{loremipsum}").unwrap();
-        assert_eq!(
-            cap,
-            (
-                "",
-                RouteParserToken::Capture(Capture::from(CaptureVariant::Named(
-                    "loremipsum".to_string()
-                )))
-            )
-        );
-    }
-
-
-    #[test]
-    fn capture_numbered_named_test() {
-        let cap = capture("{5:name}").unwrap();
-        assert_eq!(
-            cap,
-            (
-                "",
-                RouteParserToken::Capture(Capture::from(CaptureVariant::NumberedNamed {
-                    sections: 5,
-                    name: "name".to_string()
-                }))
-            )
-        );
-    }
-
-    #[test]
-    fn capture_many_named() {
-        let cap = capture("{*:name}").unwrap();
-        assert_eq!(
-            cap,
-            (
-                "",
-                RouteParserToken::Capture(Capture::from(CaptureVariant::ManyNamed(
-                    "name".to_string()
-                )))
-            )
-        );
-    }
 
     #[test]
     fn rejects_invalid_ident() {
@@ -166,8 +63,4 @@ mod test {
         valid_ident_characters("Lorem").expect("Should accept");
     }
 
-    #[test]
-    fn capture_consumes() {
-        capture("{aoeu").expect_err("Should not complete");
-    }
 }
