+++ src/matcher/mod.rs
@@ -1,7 +1,7 @@
 //! Logic for matching and capturing route strings.
 
 pub use yew_router_route_parser::{
-    parser::YewRouterParseError, Capture, CaptureVariant, Captures, MatcherToken,
+    CaptureVariant, Captures, MatcherToken,
 };
 
 mod route_matcher;
